프로필 이미지 관리 (#142)
---------

* refactor(member) : 운영 환경 -> 불필요한 채팅 요소 삭제

* refactor(task) : 불필요한 항목 제거

* refactor(chatting) : 사용하지 않는 컬럼 정리 (chat .. comment check..)

* refactor(user) : 로그인 dto 레코드 -> 클래스로 변경

* refactor(user) : 프로필 이미지 변경시 null 처리 업데이트

+++ src/main/java/com/task/weaver/common/s3/S3Uploader.java
@@ -87,10 +87,10 @@ public class S3Uploader {
     }
 
     public String updateFile(MultipartFile newFile, String oldFileName, String dirName) throws IOException {
-        // 기존 파일 삭제
-        log.info("S3 oldFileName: " + oldFileName);
-        deleteFile(oldFileName);
-        // 새 파일 업로드
+        if (oldFileName != null) {
+            log.info("S3 oldFileName: " + oldFileName);
+            deleteFile(oldFileName);
+        }
         return upload(newFile, dirName);
     }
 }
